In queue-prep, bump order status is sync in both station now even when queue is offline.
But when queue is offline and prep gets new order, after queue back online, the order prep got will not show in queue. I check on prep/expo same result, but mirror station does this perfectly. Could you apply the idea in mirror also to this? This is just like the order "ack" but right now only works in mirror. If this is a quick fix, go for it. Otherwise, work on the activation first. I will create a improvement task for later.