Try to fix the queue-expo slow issue.
We did some more testing here but seems like there is still an issue.

So we isolated the network here and set up 3 stations. We set up 2 prep stations and one queue expo.

1 prep station has auto bump set to 1 minutes
1 prep station has auto bump set to 2 minutes
1 queue expo as the queue bump set to 1 minute

We are using ID mode. We have all 4 columns enabled with the timer enabled.

We have the xml generator set for 2 seconds. Each order has 1 item on it.

After about 50-80 orders you will see that the orders on the queue expo will start slowing when the prep stations are bumping off. The timers for the orders will go up.
For example after 1 minute the order is bumped off the prep station. On the queue expo, that order will be under order received. So when that order gets to 1 minute under order received it should be moved to in preparation.
As the time goes on it takes longer and longer. So it could take 2-4 minutes for it to be bumped to prep station.

If you let it go for longer and you keep the generator running eventually the orders will stop going to the in preparation column. When you stop the generator the orders in the order received will go to the in preparation.
We have tried different android units here but all seem to do about the same thing.