Here is the issue we located with .32 version. It related to the feature named “Bumped queue order by Expeditor”.
Image

First, is this a global setting right now? we noticed when we checked or unchecked this function it would automatically apply to all stations.
Secondly, when a queue display (#7) is linked to a prep station (#6), we will not be able check or uncheck this feature on the prep station (#6); only on the queue station (#7). Ideally we would like to be able to make changes on any station regardless its expo, prep or queue. We also noticed that sometimes after assigning the primary station for the queue station, the primary station ID will not show on the Function column right away, see image below:
Image
Third, for Queue-Expo with multi-prep stations. Regardless the feature is checked or unchecked, the orders on the prep stations will not gray out first then be remove by the second bump
Lastly, would you mind making this feature unchecked by default?