One of our ISV, CBS had an issue with the XML file was picked up by the Router but not showing on the station screens. We checked every settings; they are all correct. The we realized once removed some info out of the XML it started working.
Here is the extra info that needs to be removed for the order to show in CBS’ lab:
<?xml version="1.0" encoding="UTF-8"?>
<Transaction xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">

This is not something new from CBS. I happen to have a XML sent by them last August. They had the same thing included but didn’t seem to cause any issue for them at the time.

However, when I was doing the test here I would only need to remove the highlighted part
<Transaction xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">

The router used is .31 and the KDS is .33

Would you mind checking this for us?